pyproject.toml
[project]
name = "quiz-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "NotSoSmartDev", email = "dz.sianiuk@hey.com" }]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "httpx>=0.28.1",
    "python-telegram-bot>=21.10",
]

[project.scripts]
quiz-app = "quiz_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
include = ["src"]
venvPath = "."
venv = ".venv"
